refactor: make FocusWebChromeClient a static class
